Consolidate QnAMaker dialogs
* replaced QnAMakerDialog with QnAMAkerAction code,
* Dropped QnAMakerAction class
* made QnAMAkerBuilder internal
* updated schema accordingly
* Added QnaMakerComponentRegistration() to register types
* Add .UseQnAMAker() extension for BotAdapter to register QnAMAker types
