--- libraries/Microsoft.Bot.Builder.AI.QnA/Microsoft.QnAMakerAction.schema
@@ -1,74 +1,0 @@
-{
-    "$schema": "https://raw.githubusercontent.com/microsoft/botbuilder-dotnet/4.Future/schemas/component.schema",
-    "$role": "unionType(Microsoft.IDialog)",
-    "title": "QnAMaker Action",
-    "description": "Call QnAMaker knowledge base to get an answer.",
-    "type": "object",
-    "additionalProperties": false,
-    "properties": {
-        "knowledgeBaseId": {
-            "$role": "expression",
-            "title": "KnowledgeBase Id",
-            "description": "KnowledgeBase Id of your QnA Maker KnowledgeBase.",
-            "default": "settings.qna.knowledgebaseid"
-        },
-        "endpointKey": {
-            "$role": "expression",
-            "title": "Endpoint Key",
-            "description": "Endpoint key for the QnA Maker KB.",
-            "default": "settings.qna.endpointkey"
-        },
-        "hostname": {
-            "$role": "expression",
-            "title": "Hostname",
-            "description": "Hostname for your QnA Maker service.",
-            "default": "settings.qna.hostname",
-            "examples": [
-                "https://yourserver.azurewebsites.net/qnamaker"
-            ]
-        },
-        "noAnswer": {
-            "$type": "Microsoft.IActivityTemplate",
-            "title": "Fallback answer",
-            "description": "Default answer to return when none found in KB.",
-            "default": "Sorry, I did not find an answer."
-        },
-        "threshold": {
-            "type": "number",
-            "title": "Threshold",
-            "description": "Threshold score to filter results.",
-            "default": 0.3
-        },
-        "activeLearningCardTitle": {
-            "type": "string",
-            "title": "Active learning card title",
-            "description": "Title for active learning suggestions card.",
-            "default": "Did you mean:"
-        },
-        "cardNoMatchText": {
-            "type": "string",
-            "title": "Card no match text",
-            "description": "Text for no match option.",
-            "default": "None of the above."
-        },
-        "cardNoMatchResponse ": {
-            "$type": "Microsoft.IActivityTemplate",
-            "title": "Card no match response",
-            "description": "Custom response when no match option was selected.",
-            "default": "Thanks for the feedback."
-        },
-        "strictFilters": {
-            "type": "array",
-            "title": "Strict Filter Property",
-            "description": "Memory property that holds strict filters to use when calling the QnA Maker KB.",
-            "items": {
-                "type": "object"
-            }
-        }
-    },
-    "required": [
-        "knowledgeBaseId",
-        "endpointKey",
-        "hostname"
-    ]
-}
