+++ libraries/Microsoft.Bot.Builder.AI.QnA/Microsoft.QnAMakerDialog.schema
@@ -2,53 +2,73 @@
     "$schema": "https://raw.githubusercontent.com/microsoft/botbuilder-dotnet/4.Future/schemas/component.schema",
     "$role": "unionType(Microsoft.IDialog)",
     "title": "QnAMaker Dialog",
-    "description": "This represents a dialog which is driven by a call to QnAMaker.ai knowledge base",
+    "description": "Dialog which uses QnAMAker knowledge base to answer questions.",
     "type": "object",
     "additionalProperties": false,
     "properties": {
-        "endpoint": {
-            "type": "object",
-            "title": "Endpoint",
-            "description": "This is the QnAMaker endpoint to use",
-            "properties": {
-                "knowledgeBaseId": {
-                    "type": "string",
-                    "title": "Knowledgebase Id",
-                    "description": "the knowledge base ID."
-                },
-                "endpointKey": {
-                    "type": "string",
-                    "title": "Endpoint Key",
-                    "description": "sets the endpoint key for the knowledge base"
-                },
-                "host": {
-                    "type": "string",
-                    "title": "Host",
-                    "description": "sets the host path",
-                    "examples": [
-                        "https://yourserver.azurewebsites.net/qnamaker"
-                    ]
-                }
-            },
-            "required": [
-                "knowledgeBaseId",
-                "endpointKey",
-                "host"
+        "knowledgeBaseId": {
+            "$role": "expression",
+            "title": "KnowledgeBase Id",
+            "description": "KnowledgeBase Id of your QnA Maker KnowledgeBase.",
+            "default": "settings.qna.knowledgebaseid"
+        },
+        "endpointKey": {
+            "$role": "expression",
+            "title": "Endpoint Key",
+            "description": "Endpoint key for the QnA Maker KB.",
+            "default": "settings.qna.endpointkey"
+        },
+        "hostname": {
+            "$role": "expression",
+            "title": "Hostname",
+            "description": "Hostname for your QnA Maker service.",
+            "default": "settings.qna.hostname",
+            "examples": [
+                "https://yourserver.azurewebsites.net/qnamaker"
             ]
         },
-        "options": {
-            "type": "object",
-            "title": "Options",
-            "properties": {
-                "scoreThreshold": {
-                    "type": "number",
-                    "title": "Score Threshold",
-                    "description": "\"sets the minimum score threshold, used to filter returned results. Scores are normalized to the range of 0.0 to 1.0"
-                }
+        "noAnswer": {
+            "$type": "Microsoft.IActivityTemplate",
+            "title": "Fallback answer",
+            "description": "Default answer to return when none found in KB.",
+            "default": "Sorry, I did not find an answer."
+        },
+        "threshold": {
+            "type": "number",
+            "title": "Threshold",
+            "description": "Threshold score to filter results.",
+            "default": 0.3
+        },
+        "activeLearningCardTitle": {
+            "type": "string",
+            "title": "Active learning card title",
+            "description": "Title for active learning suggestions card.",
+            "default": "Did you mean:"
+        },
+        "cardNoMatchText": {
+            "type": "string",
+            "title": "Card no match text",
+            "description": "Text for no match option.",
+            "default": "None of the above."
+        },
+        "cardNoMatchResponse ": {
+            "$type": "Microsoft.IActivityTemplate",
+            "title": "Card no match response",
+            "description": "Custom response when no match option was selected.",
+            "default": "Thanks for the feedback."
+        },
+        "strictFilters": {
+            "type": "array",
+            "title": "Strict Filter Property",
+            "description": "Memory property that holds strict filters to use when calling the QnA Maker KB.",
+            "items": {
+                "type": "object"
             }
         }
     },
     "required": [
-        "endpoint"
+        "knowledgeBaseId",
+        "endpointKey",
+        "hostname"
     ]
 }
