+++ libraries/Microsoft.Bot.Builder.AI.QnA/QnAMakerActionBuilder.cs
@@ -15,26 +15,26 @@ namespace Microsoft.Bot.Builder.AI.QnA
     /// <summary>
     /// QnAMaker action builder class
     /// </summary>
-    public class QnAMakerActionBuilder
+    internal class QnAMakerActionBuilder
     {
         /// <summary>
         /// QnA Maker action builder
         /// </summary>
-        public const string QnAMakerDialogName = "qnamaker-dialog";
+        internal const string QnAMakerDialogName = "qnamaker-dialog";
 
         // Dialog Options parameters
-        public const float DefaultThreshold = 0.3F;
-        public const int DefaultTopN = 3;
-        public const string DefaultNoAnswer = "No QnAMaker answers found.";
+        internal const float DefaultThreshold = 0.3F;
+        internal const int DefaultTopN = 3;
+        internal const string DefaultNoAnswer = "No QnAMaker answers found.";
 
         // Card parameters
-        public const string DefaultCardTitle = "Did you mean:";
-        public const string DefaultCardNoMatchText = "None of the above.";
-        public const string DefaultCardNoMatchResponse = "Thanks for the feedback.";
+        internal const string DefaultCardTitle = "Did you mean:";
+        internal const string DefaultCardNoMatchText = "None of the above.";
+        internal const string DefaultCardNoMatchResponse = "Thanks for the feedback.";
 
         // Define value names for values tracked inside the dialogs.
-        public const string QnAOptions = "qnaOptions";
-        public const string QnADialogResponseOptions = "qnaDialogResponseOptions";
+        internal const string QnAOptions = "qnaOptions";
+        internal const string QnADialogResponseOptions = "qnaDialogResponseOptions";
         private const string CurrentQuery = "currentQuery";
         private const string QnAData = "qnaData";
         private const string QnAContextData = "qnaContextData";
@@ -52,7 +52,7 @@ namespace Microsoft.Bot.Builder.AI.QnA
         /// Dialog helper to generate dialogs.
         /// </summary>
         /// <param name="services">Bot Services.</param>
-        public QnAMakerActionBuilder(QnAMaker services)
+        internal QnAMakerActionBuilder(QnAMaker services)
         {
             _qnaMakerDialog = new WaterfallDialog(QnAMakerDialogName)
                 .AddStep(CallGenerateAnswerAsync)
@@ -67,7 +67,7 @@ namespace Microsoft.Bot.Builder.AI.QnA
         /// </summary>
         /// <param name="dc">DialogContext</param>
         /// <returns>Updated dialog context.</returns>
-        public DialogContext BuildDialog(DialogContext dc)
+        internal DialogContext BuildDialog(DialogContext dc)
         {
             if (dc == null)
             {
