+++ schemas/sdk.schema
@@ -326,13 +326,8 @@
             "$ref": "#/definitions/Microsoft.PhoneNumberEntityRecognizer"
         },
         {
-            "title": "Microsoft.QnAMakerAction",
-            "description": "Call QnAMaker knowledge base to get an answer.",
-            "$ref": "#/definitions/Microsoft.QnAMakerAction"
-        },
-        {
             "title": "Microsoft.QnAMakerDialog",
-            "description": "This represents a dialog which is driven by a call to QnAMaker.ai knowledge base",
+            "description": "Dialog which uses QnAMAker knowledge base to answer questions.",
             "$ref": "#/definitions/Microsoft.QnAMakerDialog"
         },
         {
@@ -3027,13 +3022,8 @@
                     "$ref": "#/definitions/Microsoft.OAuthInput"
                 },
                 {
-                    "title": "Microsoft.QnAMakerAction",
-                    "description": "Call QnAMaker knowledge base to get an answer.",
-                    "$ref": "#/definitions/Microsoft.QnAMakerAction"
-                },
-                {
                     "title": "Microsoft.QnAMakerDialog",
-                    "description": "This represents a dialog which is driven by a call to QnAMaker.ai knowledge base",
+                    "description": "Dialog which uses QnAMAker knowledge base to answer questions.",
                     "$ref": "#/definitions/Microsoft.QnAMakerDialog"
                 },
                 {
@@ -5865,10 +5855,10 @@
                 }
             ]
         },
-        "Microsoft.QnAMakerAction": {
+        "Microsoft.QnAMakerDialog": {
             "$role": "unionType(Microsoft.IDialog)",
-            "title": "QnAMaker Action",
-            "description": "Call QnAMaker knowledge base to get an answer.",
+            "title": "QnAMaker Dialog",
+            "description": "Dialog which uses QnAMAker knowledge base to answer questions.",
             "type": "object",
             "additionalProperties": false,
             "properties": {
@@ -5877,7 +5867,7 @@
                     "description": "Defines the valid properties for the component you are configuring (from a dialog .schema file)",
                     "type": "string",
                     "pattern": "^[a-zA-Z][a-zA-Z0-9.]*$",
-                    "const": "Microsoft.QnAMakerAction"
+                    "const": "Microsoft.QnAMakerDialog"
                 },
                 "$copy": {
                     "title": "$copy",
@@ -5979,100 +5969,6 @@
                         "knowledgeBaseId",
                         "endpointKey",
                         "hostname",
-                        "$type"
-                    ]
-                }
-            ]
-        },
-        "Microsoft.QnAMakerDialog": {
-            "$role": "unionType(Microsoft.IDialog)",
-            "title": "QnAMaker Dialog",
-            "description": "This represents a dialog which is driven by a call to QnAMaker.ai knowledge base",
-            "type": "object",
-            "additionalProperties": false,
-            "properties": {
-                "$type": {
-                    "title": "$type",
-                    "description": "Defines the valid properties for the component you are configuring (from a dialog .schema file)",
-                    "type": "string",
-                    "pattern": "^[a-zA-Z][a-zA-Z0-9.]*$",
-                    "const": "Microsoft.QnAMakerDialog"
-                },
-                "$copy": {
-                    "title": "$copy",
-                    "description": "Copy the definition by id from a .dialog file.",
-                    "type": "string",
-                    "pattern": "^(([a-zA-Z][a-zA-Z0-9.]*)?(#[a-zA-Z][a-zA-Z0-9.]*)?)$"
-                },
-                "$id": {
-                    "title": "$id",
-                    "description": "Inline id for reuse of an inline definition",
-                    "type": "string",
-                    "pattern": "^([a-zA-Z][a-zA-Z0-9.]*)$"
-                },
-                "$designer": {
-                    "title": "$designer",
-                    "type": "object",
-                    "description": "Extra information for the Bot Framework Designer."
-                },
-                "endpoint": {
-                    "type": "object",
-                    "title": "Endpoint",
-                    "description": "This is the QnAMaker endpoint to use",
-                    "required": [
-                        "knowledgeBaseId",
-                        "endpointKey",
-                        "host"
-                    ],
-                    "properties": {
-                        "knowledgeBaseId": {
-                            "type": "string",
-                            "title": "Knowledgebase Id",
-                            "description": "the knowledge base ID."
-                        },
-                        "endpointKey": {
-                            "type": "string",
-                            "title": "Endpoint Key",
-                            "description": "sets the endpoint key for the knowledge base"
-                        },
-                        "host": {
-                            "type": "string",
-                            "title": "Host",
-                            "description": "sets the host path",
-                            "examples": [
-                                "https://yourserver.azurewebsites.net/qnamaker"
-                            ]
-                        }
-                    }
-                },
-                "options": {
-                    "type": "object",
-                    "title": "Options",
-                    "properties": {
-                        "scoreThreshold": {
-                            "type": "number",
-                            "title": "Score Threshold",
-                            "description": "\"sets the minimum score threshold, used to filter returned results. Scores are normalized to the range of 0.0 to 1.0"
-                        }
-                    }
-                }
-            },
-            "patternProperties": {
-                "^\\$": {
-                    "type": "string"
-                }
-            },
-            "anyOf": [
-                {
-                    "title": "Reference",
-                    "required": [
-                        "$copy"
-                    ]
-                },
-                {
-                    "title": "Type",
-                    "required": [
-                        "endpoint",
                         "$type"
                     ]
                 }
